refactor(fsd): Step 3 — entities layer migration
entities/voc, user, notification, master 레이어 신설.
각 도메인 api/model/ui 를 entities/ 로 이동하고
구 경로(src/api/, src/lib/, src/components/voc/)는
@entities/* 별칭 re-export stub 으로 대체.
Step 7 cleanup 전까지 기존 callers 무변경 동작 보장.

- typecheck clean / 374 tests pass
- entities → features 역방향 import 0건
