+++ frontend/src/components/voc/VocStatusBadge.tsx
@@ -1,15 +1,2 @@
-import type { VocStatus } from '../../../../shared/contracts/voc';
-import { VOC_STATUS_SLUG } from '../../lib/voc-status-slug';
-import { SolidChip } from '../../shared/ui/badge';
-
-export function VocStatusBadge({ status }: { status: VocStatus }) {
-  const slug = VOC_STATUS_SLUG[status];
-  return (
-    <SolidChip
-      variant={slug}
-      label={status}
-      extraTestId={`status-badge-${status}`}
-      ariaLabelOverride={`상태 ${status}`}
-    />
-  );
-}
+// re-export stub — real implementation moved to entities/voc/ui/VocStatusBadge (Step 3)
+export { VocStatusBadge } from '@entities/voc/ui/VocStatusBadge';
