+++ frontend/src/entities/master/index.ts
@@ -1,0 +1,1 @@
+export { mastersApi } from './api/masterApi';
