+++ frontend/src/entities/notification/api/notificationQueryKeys.ts
@@ -1,0 +1,6 @@
+import type { Role } from '@contracts/common';
+
+export const notificationQueryKeys = {
+  all: (role: Role) => ['notifications', role] as const,
+  list: (role: Role) => ['notifications', role, 'list'] as const,
+} as const;
