+++ frontend/src/entities/notification/index.ts
@@ -1,0 +1,2 @@
+export { notificationsApi } from './api/notificationApi';
+export { notificationQueryKeys } from './api/notificationQueryKeys';
