+++ frontend/src/entities/user/index.ts
@@ -1,0 +1,3 @@
+export { getMe } from './api/userApi';
+export type { AuthUser, AuthRole } from './api/userApi';
+export { userQueryKeys } from './api/userQueryKeys';
