+++ frontend/src/entities/voc/index.ts
@@ -1,0 +1,11 @@
+export { vocApi } from './api/vocApi';
+export { vocQueryKeys } from './api/vocQueryKeys';
+export { VOC_STATUS_SLUG } from './model/vocStatus';
+export { getVocTypeIconConfig } from './model/vocType';
+export type { VocTypeIconConfig } from './model/vocType';
+export { VocStatusBadge } from './ui/VocStatusBadge';
+export { VocPriorityBadge } from './ui/VocPriorityBadge';
+export { VocTagPill } from './ui/VocTagPill';
+export type { VocTagPillProps } from './ui/VocTagPill';
+export { VocTypeBadge } from './ui/VocTypeBadge';
+export type { VocTypeBadgeProps } from './ui/VocTypeBadge';
